backend/requirements.txt
alembic==1.14.1
celery[redis]==5.4.0
fastapi==0.115.6
pydantic-settings==2.7.0
PyJWT==2.10.1
psycopg[binary]==3.2.13
python-multipart==0.0.19
SQLAlchemy==2.0.36
uvicorn[standard]==0.32.1

# test dependencies
httpx==0.28.1
pytest==8.3.4
